removed alter-users migration, users.balance changed to integer, added Dinero to payout transaction

+++ migrations/sqls/20230705162226-create-balance-history-up.sql
@@ -1,8 +1,8 @@
 CREATE TABLE `balance_history` (
   `id` INT NOT NULL AUTO_INCREMENT,
   `user_id` INT NOT NULL,
-  `prev_balance` DECIMAL(10, 2) NOT NULL,
-  `new_balance` DECIMAL(10, 2) NOT NULL,
+  `prev_balance` INT NOT NULL,
+  `new_balance` INT NOT NULL,
   `operation` VARCHAR(255) NOT NULL,
   `extra` VARCHAR(255),
   `date` TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
